Move TextFieldAssist.Hint&HintOpacity into HintAssist
