+++ MaterialDesignThemes.Wpf/Themes/MaterialDesignTheme.TextBox.xaml
@@ -78,10 +78,10 @@
                                               UseLayoutRounding="{TemplateBinding UseLayoutRounding}"
                                               />
                                 <wpf:SmartHint x:Name="Hint"
-                                               Hint="{Binding Path=(wpf:TextFieldAssist.Hint), RelativeSource={RelativeSource TemplatedParent}}"
+                                               Hint="{Binding Path=(wpf:HintAssist.Hint), RelativeSource={RelativeSource TemplatedParent}}"
                                                HintProxy="{Binding RelativeSource={RelativeSource TemplatedParent}, Converter={x:Static converters:HintProxyFabricConverter.Instance}}"
                                                FontSize="{TemplateBinding FontSize}"
-                                               HintOpacity="{Binding Path=(wpf:TextFieldAssist.HintOpacity), RelativeSource={RelativeSource TemplatedParent}}"
+                                               HintOpacity="{Binding Path=(wpf:HintAssist.HintOpacity), RelativeSource={RelativeSource TemplatedParent}}"
                                                UseFloating="{Binding Path=(wpf:HintAssist.UseFloating), RelativeSource={RelativeSource TemplatedParent}}"
                                                />
                             </Grid>
