+++ MaterialDesignThemes.Wpf/Themes/MaterialDesignTheme.TimePicker.xaml
@@ -14,7 +14,7 @@
 		<Setter Property="HorizontalContentAlignment" Value="Stretch"/>
         <Setter Property="IsTabStop" Value="False"/>
         <Setter Property="Margin" Value="0 12 0 0" />
-        <Setter Property="wpf:TextFieldAssist.Hint" Value="Select time" />
+        <Setter Property="wpf:HintAssist.Hint" Value="Select time" />
 		<Setter Property="ClockStyle" Value="{DynamicResource MaterialDesignClock}" />
 		<Setter Property="ClockHostContentControlStyle">
 			<Setter.Value>
@@ -51,9 +51,9 @@
                                 <wpf:SmartHint x:Name="Hint"
                                                FontSize="{TemplateBinding FontSize}"
                                                UseFloating="{Binding Path=(wpf:HintAssist.UseFloating), RelativeSource={RelativeSource TemplatedParent}}"
-                                               Hint="{Binding Path=(wpf:TextFieldAssist.Hint), RelativeSource={RelativeSource TemplatedParent}}"
+                                               Hint="{Binding Path=(wpf:HintAssist.Hint), RelativeSource={RelativeSource TemplatedParent}}"
                                                HintProxy="{Binding RelativeSource={RelativeSource TemplatedParent}, Converter={x:Static converters:HintProxyFabricConverter.Instance}}"
-                                               HintOpacity="{Binding Path=(wpf:TextFieldAssist.HintOpacity), RelativeSource={RelativeSource TemplatedParent}}"/>
+                                               HintOpacity="{Binding Path=(wpf:HintAssist.HintOpacity), RelativeSource={RelativeSource TemplatedParent}}"/>
 							</Grid>
 						</ControlTemplate>
 						<ControlTemplate x:Key="DropDownButtonTemplate" TargetType="{x:Type Button}">
